backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djangorestframework_simplejwt == 5.5.0
djoser==2.1.0
Pillow==9.3.0
psycopg2==2.8.6
pybase64==1.4.1
pytest==6.2.4
pytest-django==4.4.0
python-dotenv==1.0.1
pytest-pythonpath==0.7.3
PyYAML==6.0